test: move copyAllFiles to a static method to be reused
