+++ zeebe/snapshot/src/test/java/io/camunda/zeebe/snapshots/impl/FileBasedSnapshotStoreTest.java
@@ -7,6 +7,7 @@
  */
 package io.camunda.zeebe.snapshots.impl;
 
+import static io.camunda.zeebe.snapshots.SnapshotCopyUtil.copyAllFiles;
 import static java.nio.file.StandardOpenOption.CREATE_NEW;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
@@ -25,12 +26,10 @@ import java.io.UncheckedIOException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.StandardCopyOption;
 import java.nio.file.StandardOpenOption;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.function.BiConsumer;
 import java.util.zip.CRC32C;
 import org.junit.Before;
 import org.junit.Rule;
@@ -552,27 +551,6 @@ public class FileBasedSnapshotStoreTest {
             })
         .join();
     return transientSnapshot;
-  }
-
-  private BiConsumer<Path, Path> copyAllFiles() {
-    return (source, target) -> {
-      try (final var stream = Files.walk(source)) {
-        stream.forEach(
-            path -> {
-              if (!source.equals(path)) {
-                try {
-                  final var relativePath = source.relativize(path);
-                  final var targetPath = target.resolve(relativePath);
-                  Files.copy(path, targetPath, StandardCopyOption.REPLACE_EXISTING);
-                } catch (final IOException e) {
-                  throw new UncheckedIOException(e);
-                }
-              }
-            });
-      } catch (final IOException e) {
-        throw new RuntimeException(e);
-      }
-    };
   }
 
   private TransientSnapshot takeTransientSnapshot(
